Backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-dotenv==1.0.0
supabase==1.0.3
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
